feat: single page for group 11 charts
