+++ src/pages/Group11/DeathRate.tsx
@@ -2,9 +2,6 @@ import { StackedBarChart } from '@carbon/charts-react';
 import '@carbon/charts/styles.css';
 import "@carbon/styles/css/styles.css";
 import { ScaleTypes } from '@carbon/charts/interfaces';
-import { useLoaderData } from 'react-router-dom';
-import { stackedBarLoader } from '../loaders';
-import { LoaderData } from '../types';
 
 const options = {
 	title: 'Daily Covid Cases and Deaths in Each State',
@@ -23,19 +20,18 @@ const options = {
 	height: '400px',
 };
 
-function StackedBarChartPage () {
-	const loaderData = useLoaderData() as LoaderData<typeof stackedBarLoader>;
-	let data: any[] = [];
-	loaderData.forEach((item: any) => {
-		data.push(...item);
+function DeathRate ({ data }: { data: any }) {
+	let processed_data: any[] = [];
+	data.forEach((item: any) => {
+		processed_data.push(...item);
 	})
 
 	return (
-		<div style={{ width: '100%', height: '400px' }}>
+		<div>
 			<h1>Stacked Bar Chart</h1>
-			<StackedBarChart data={data} options={options} />
+			<StackedBarChart data={processed_data} options={options} />
 		</div>
 	);
 }
 
-export default StackedBarChartPage;
+export default DeathRate;
