+++ src/pages/Group11/VaccinationRate.tsx
@@ -2,9 +2,6 @@ import BulletChart from "@carbon/charts-react/bullet-chart";
 import { ScaleTypes } from "@carbon/charts/interfaces";
 import "@carbon/charts/styles.css";
 import "@carbon/styles/css/styles.css";
-import { useLoaderData } from "react-router-dom";
-import { vacRateLoader } from "../loaders";
-import { LoaderData } from "../types";
 
 const options = {
     "title": "Percentage of Vaccinated People",
@@ -12,7 +9,7 @@ const options = {
         "bottom": {
             "mapsTo": "value",
             "extendLinearDomainBy": "max"
-        }, 
+        },
         "left": {
             "scaleType": ScaleTypes.LABELS,
             "mapsTo": "title"
@@ -25,14 +22,13 @@ const options = {
     "height": "400px"
 };
 
-const Vaccine = () => {
-    const data = useLoaderData() as LoaderData<typeof vacRateLoader>;
+const VaccinationRate = ({ data }: { data: any }) => {
     return (
         <div>
-            <h1>my covid view</h1>
+            <h1>Bullet Chart</h1>
             <BulletChart data={data} options={options} />
         </div>
     );
 };
 
-export default Vaccine;
+export default VaccinationRate;
