+++ src/pages/index.ts
@@ -2,5 +2,4 @@ export { default as Overview } from "./Overview";
 export { default as Feedback } from "./Feedback";
 export { default as SunburstPage } from "./SunburstPage";
 export { default as TreemapPage } from "./TreeMapPage";
-export { default as StackedBarChartPage } from "./StackedBarChartPage";
-export { default as VaccinationRate } from "./VaccinationRate";
+export { default as Group11Page } from "./Group11";
